requirements.txt
snowflake-connector-python>=3.6.0

